fix(util): handle goccy/go-yaml uint64 type for numeric settings
goccy/go-yaml v1.19.2 parses non-negative YAML integers as uint64 and
negative integers as int64, not int/float64. The original GetIntSetting
only checked int and float64, causing all YAML integer settings to
silently fall back to defaults:

- Modbus port 802 → silently reverted to 502
- S7 port/rack/slot → silently reverted to defaults
- OPC UA subscription-buffer/max-batch-size → silently reverted
- MQTT QoS 2 → silently downgraded to 1
- Parser tag-from-topic → silently reverted to 0 (tag never extracted)
- S7 TypeFloat64 write → silently wrote 0.0 instead of actual value

Fixed in 4 locations:
- common/util.GetIntSetting: switch type covering int/int64/uint64/float64
- transport/parser.getIntSetting: removed duplicate, now delegates to util.GetIntSetting
- driver/opcua subscription-buffer: now uses util.GetIntSetting with >0 guard
- driver/s7 TypeFloat64 write: now uses util.ToFloat64 for type-safe conversion

Tests updated to use uint64 (matching goccy/go-yaml output) instead of
Go int literals that masked the bug.

+++ driver/opcua/client_test.go
@@ -564,7 +564,7 @@ func TestOPCUASubscriptionConfig(t *testing.T) {
 
 	t.Run("custom subscription buffer size", func(t *testing.T) {
 		cfg := validConfig("x")
-		cfg.Settings["subscription-buffer"] = 4096
+		cfg.Settings["subscription-buffer"] = uint64(4096)
 		drv, err := NewOPCUADriver(cfg)
 		if err != nil {
 			t.Fatalf("NewOPCUADriver: %v", err)
